Add files via upload
The code provided asks for the user input of two MLB teams and produces a winner of a theoretical game between the two using data from the 2022 and 2021 seasons. All of the data needed to run the program is provided within the wrapper, so there is no need to reference an outside database in a different directory. Just download the R file and call the function to use the program! The predictive model within uses OPS+, ERA+, Walks Allowed, Winning percentage in one-run games, Percentage of runs scored, Number of total all-star appearances of the team’s players, Homeruns allowed, Number of strikeouts pitching, Homeruns hit, The team rank after the 2021 season, and Fielding percentage to produce a predicted ranking for each team. The code will find each team-provided ranking and compare it to two to predict who will win, with a tad bit of randomness involved. To start, the function will ask a user to input the home team first and the away team after, continually prompting the user until they type in the full team name used in the database. A list of the available teams is given before being prompted to enter. The code will also continue prompting for a second team if the same team is entered twice. Then the winner will be provided based on the team’s predicted ranking!